+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
+- CLI: `bones_api` with `serve` command.
 - Initial version.